api-backend/requirements.txt
flask
flask_cors
python-dotenv
google-generativeai
gunicorn
